provisions the new instance
Uses cloud init to install fluent bit and configure it, generating
random events to ingest. It doesn't go much of anywhere, and the
counter output seems unreliable, but manually adding a local file
output shows that the events are flowing. Next up is to add a second
instance with an http input to forward events to.